refactor: create separate participant more dropdown
